% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMSM.R
\name{BMSM}
\alias{BMSM}
\title{Bayesian Multi-scale Model for Smoothing Poisson sequence}
\usage{
BMSM(x, sc = 1, bmsm_control = list())
}
\arguments{
\item{x:}{a vector of observed counts}

\item{sc:}{a scale factor/offset, must be a scalar}

\item{bmsm_control:}{control parameters of BMSM, see bmsm_control_default()}
}
\value{
estimated poisson sequence
}
\description{
Bayesian Multi-scale Model for Smoothing Poisson sequence
}
