% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_smooth_fn.R
\name{dwt.fn}
\alias{dwt.fn}
\title{Empirical Bayes wavelet smoothing for flashier}
\usage{
dwt.fn(x, s, g_init, fix_g, output)
}
\arguments{
\item{x}{data}

\item{s}{known standard error}

\item{filter.number, family}{wavelet family and filter number as in `wavethresh` package}

\item{ebnm_params}{a list of `ebnm` parameters}

\item{W}{the dwt matrix for calc posterior variance. Remove the first row which is all 1/sqrt(n).}
}
\value{
a list of
 \item{mu.est:}{posterior mean}
 \item{mu.var:}{posterior variance}
 \item{loglik:}{log likelihood}
 \item{dKL:}{KL divergence between g(the prior) and q(the posterior)}
}
\description{
Smooth homogeneous Gaussian data.
}
