% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebnm.R
\name{ebnm_normal}
\alias{ebnm_normal}
\title{A function to do eb for normal mean}
\usage{
ebnm_normal(x, s, eps = 1e-08)
}
\arguments{
\item{x:}{a vector of data}

\item{s:}{a vector of sd}

\item{eps:}{adjustment}
}
\value{
eb estimate of prior mu, sd and posterior mean
}
\description{
A function to do eb for normal mean
}
