% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgp.R
\name{ebnm_sgp}
\alias{ebnm_sgp}
\title{Sparse Gaussian process in EBNM syntax}
\usage{
ebnm_sgp(x, s, g_init = NULL, fix_g = FALSE, output)
}
\arguments{
\item{x}{A vector of observations.}

\item{s}{A vector of standard errors}

\item{g_init}{a list of X_ind,kernel_param,kernel_func,mu}

\item{fix_g}{If TRUE, the prior is fixed at g_init instead of estimated}

\item{output}{ignore this.}
}
\value{
a list of posterior, log-likelihood, fitted_g
}
\description{
Sparse Gaussian process in EBNM syntax
}
