% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_sgp.R
\name{ebpm_pois_sgp}
\alias{ebpm_pois_sgp}
\title{Poisson Sparse Gaussian Process}
\usage{
ebpm_pois_sgp(
  x,
  s = 1,
  g_init = NULL,
  fix_g = FALSE,
  q_init = NULL,
  control = list()
)
}
\arguments{
\item{x}{A vector of Poisson observations.}

\item{s}{A vector of scaling factors for Poisson observations}

\item{g_init}{a list of prior parameters: X_ind, kernel_param, mu, kernel_func}

\item{fix_g}{a boolean or a vector of boolean of length 3. If TRUE, set the corresponding prior at g_init.}

\item{q_init}{a list of init values of posterior: mean_log_ind, v_log_ind. These are the posterior mean and var of the f at the inducing points.}

\item{control}{a list of other parameters passed to the `pois_sgp` function.}
}
\value{
a list of posterior, log_likelihood, fitted_g
}
\description{
Poisson Sparse Gaussian Process
}
\details{
The model is
\deqn{x_{i}\sim \text{Poisson}(s_i\exp(\mu+f_{i})),}
\deqn{f\sim N(0,\Sigma(\theta)).}
}
