% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebps.R
\name{ebps}
\alias{ebps}
\title{empirical Bayes Poisson smoothing}
\usage{
ebps(
  x,
  s = NULL,
  g_init = NULL,
  q_init = NULL,
  init_control = list(),
  general_control = list(),
  smooth_control = list()
)
}
\arguments{
\item{x, s}{data vector and scaling factor}

\item{g_init}{a list of sigma2, and g_smooth.}

\item{q_init}{a list of m, smooth. m is the posterior mean of mu, smooth the posteriro mean of b}

\item{init_control}{See function ebps_init_control_default}

\item{general_control}{See function ebps_general_control_default}

\item{smooth_control}{See function ebps_smooth_control_default}
}
\description{
empirical Bayes Poisson smoothing
}
\details{
The problem is
\deqn{x_i\sim Poisson(\lambda_i,}
\deqn{\lambda_i = \exp(\mu_i)),}
\deqn{\mu_i\sim N(b_i,\sigma^2),}
\deqn{\b_i\sim g(.).}
}
\examples{
set.seed(12345)
n=2^9
sigma=0.5
mu=c(rep(0.3,n/4), rep(3, n/4), rep(10, n/4), rep(0.3, n/4))
x = rpois(n,exp(log(mu)+rnorm(n,sd=sigma)))
fit = ebps(x)
plot(x,col='grey80')
lines(fit$posterior$mean_smooth)
fit$sigma2
plot(fit$elbo_trace)
}
