% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebps_init.R
\name{ebps_init}
\alias{ebps_init}
\title{init ebps}
\usage{
ebps_init(x, s, make_power_of_2, g_init, q_init, m_init_method)
}
\arguments{
\item{g_init}{a list of sigma2, and g_smooth}

\item{q_init}{a list of m, smooth}
}
\description{
init ebps
}
