% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebps_twostep.R
\name{ebps_twostep}
\alias{ebps_twostep}
\title{Apply ebps then run smash as a two-step procedure}
\usage{
ebps_twostep(
  x,
  s = NULL,
  g_init = NULL,
  q_init = NULL,
  init_control = list(),
  general_control = list(),
  smooth_control = list(),
  homo = FALSE
)
}
\arguments{
\item{x, s}{data vector and scaling factor. s can be a vector of the same length as x, or a scalar.}

\item{g_init}{a list of initial value of sigma2, and g_smooth. g_smooth is the initial prior g of the smoothing method. Can be NULL.}

\item{q_init}{a list of initial values of m, smooth. m is the posterior mean of mu, smooth the posterior mean of b. See the details below.}

\item{init_control}{See function ebps_init_control_default}

\item{general_control}{See function ebps_general_control_default}

\item{smooth_control}{See function ebps_smooth_control_default}
}
\description{
Apply ebps then run smash as a two-step procedure
}
