% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebtf_mr_ash.R
\name{ebtf_mr_ash}
\alias{ebtf_mr_ash}
\title{Empirical bayes Trend filtering via mr ash}
\usage{
ebtf_mr_ash(x, k)
}
\arguments{
\item{x}{observed Poisson sequence}
}
\description{
Empirical bayes Trend filtering via mr ash
}
