% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fine_coeff_rm.R
\name{fine_coeff_rm}
\alias{fine_coeff_rm}
\title{A function that sets the finest level coefficience to zero}
\usage{
fine_coeff_rm(y0, family, filter.number)
}
\arguments{
\item{:}{y0}
}
\value{
wavelet
}
\description{
A function that sets the finest level coefficience to zero
}
