% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_sgp.R
\name{grad_kernel_param}
\alias{grad_kernel_param}
\title{Get derivative of log kernel parameters and mean}
\usage{
grad_kernel_param(
  kernel_param,
  X,
  y,
  sc,
  X_ind,
  kernel_func,
  post_mean,
  V,
  mu,
  Jitter,
  d_mat_nm,
  d_mat_mm
)
}
\description{
Get derivative of log kernel parameters and mean
}
