% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_sgp.R
\name{grad_post_mean}
\alias{grad_post_mean}
\title{Gradient and hessian of eblo w.r.t post_mean}
\usage{
grad_post_mean(
  X,
  y,
  sc,
  X_ind,
  kernel_param,
  kernel_func,
  post_mean,
  V,
  mu,
  return_grad = T,
  return_hess = T,
  Jitter
)
}
\description{
Gradient and hessian of eblo w.r.t post_mean
}
