% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgp.R
\name{Maternkernel}
\alias{Maternkernel}
\title{Matern Kernel}
\usage{
Maternkernel(
  X1,
  X2,
  kernel_param,
  eps = NULL,
  diagonal = F,
  nu = 3/2,
  d_mat = NULL,
  calc_grad_scales = FALSE
)
}
\description{
Matern Kernel
}
