% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgp_torch.R
\name{Maternkernel_torch}
\alias{Maternkernel_torch}
\title{Matern Kernel}
\usage{
Maternkernel_torch(
  X1,
  X2,
  kernel_param,
  eps = NULL,
  diagonal = F,
  nu = 3/2,
  d_mat = NULL
)
}
\arguments{
\item{d_mat}{this is x_i - x_j}
}
\description{
Matern Kernel
}
