% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_gp_split.R
\name{neg_log_likelihood}
\alias{neg_log_likelihood}
\title{Define the negative marginal log-likelihood}
\usage{
neg_log_likelihood(params, x, s2, grid)
}
\description{
Define the negative marginal log-likelihood
}
