% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_gp_split.R
\name{neg_log_likelihood_gradient}
\alias{neg_log_likelihood_gradient}
\title{Define the gradient of the negative log-likelihood}
\usage{
neg_log_likelihood_gradient(params, x, s2, grid)
}
\description{
Define the gradient of the negative log-likelihood
}
