% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_reg_split.R
\name{pois_reg_splitting}
\alias{pois_reg_splitting}
\title{this function fits empirical Bayes Sparse Poisson regression}
\usage{
pois_reg_splitting(
  X,
  y,
  fix_g = FALSE,
  g_init = NULL,
  maxiter = 100,
  maxiter_mrash = 100,
  printevery = 10,
  tol = 1e-05
)
}
\description{
this function fits empirical Bayes Sparse Poisson regression
}
