% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_sgp.R
\name{pois_sgp_matrix_helper}
\alias{pois_sgp_matrix_helper}
\title{Helper function for matrix calculations}
\usage{
pois_sgp_matrix_helper(
  X,
  y,
  sc,
  X_ind,
  kernel_param,
  kernel_func,
  post_mean,
  V,
  mu,
  Jitter,
  get_delta = T,
  d_mat_nm,
  d_mat_mm
)
}
\description{
Helper function for matrix calculations
}
