% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_sgp.R
\name{pois_sgp_obj_for_optim}
\alias{pois_sgp_obj_for_optim}
\title{objective function for optimizing elbo w.r.t X-ind and kernel_param for `optim`}
\usage{
pois_sgp_obj_for_optim(
  params,
  m,
  n_kernel_param,
  X,
  y,
  sc,
  X_ind,
  kernel_param,
  kernel_func,
  post_mean,
  V,
  mu,
  fix_X_ind,
  fix_kernel_param,
  Jitter,
  d_mat_nm,
  d_mat_mm
)
}
\description{
objective function for optimizing elbo w.r.t X-ind and kernel_param for `optim`
}
