% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_sgp.R
\name{pois_sgp_obj_for_optim_kernel_only}
\alias{pois_sgp_obj_for_optim_kernel_only}
\title{objective function for optimizing elbo w.r.t kernel_param for `optim`}
\usage{
pois_sgp_obj_for_optim_kernel_only(
  kernel_param,
  X,
  y,
  sc,
  X_ind,
  kernel_func,
  post_mean,
  V,
  mu,
  Jitter
)
}
\description{
objective function for optimizing elbo w.r.t kernel_param for `optim`
}
