% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_sgp.R
\name{pois_sgp_optimize_kernel_newton}
\alias{pois_sgp_optimize_kernel_newton}
\title{Newton's method for optimizing elbo w.r.t kernel_param, using numerical grad and hess}
\usage{
pois_sgp_optimize_kernel_newton(
  kernel_param,
  X,
  y,
  sc,
  X_ind,
  kernel_func,
  post_mean,
  V,
  mu,
  Jitter,
  d_mat_nm,
  d_mat_mm,
  tol = 1e-05,
  maxiter = 100
)
}
\description{
Newton's method for optimizing elbo w.r.t kernel_param, using numerical grad and hess
}
