% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_sgp.R
\name{pois_sgp_update_mu}
\alias{pois_sgp_update_mu}
\title{Function for optimize mu}
\usage{
pois_sgp_update_mu(
  X,
  y,
  sc,
  X_ind,
  kernel_param,
  kernel_func,
  post_mean,
  V,
  mu,
  fix_mu,
  Jitter
)
}
\description{
Function for optimize mu
}
