% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_sgp.R
\name{pois_sgp_update_post_mean}
\alias{pois_sgp_update_post_mean}
\title{Function for optimizing posterior mean}
\usage{
pois_sgp_update_post_mean(
  X,
  y,
  sc,
  X_ind,
  kernel_param,
  kernel_func,
  post_mean,
  V,
  mu,
  maxiter = 100,
  tol = 1e-05,
  Jitter,
  d_mat_nm,
  d_mat_mm,
  verbose
)
}
\description{
Function for optimizing posterior mean
}
