% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_sgp.R
\name{pois_sgp_update_prior}
\alias{pois_sgp_update_prior}
\title{Master function for optimizing priors (X_ind, kernel_param)}
\usage{
pois_sgp_update_prior(
  X,
  y,
  sc,
  X_ind,
  kernel_param,
  kernel_func,
  post_mean,
  V,
  mu,
  fix_kernel_param,
  fix_X_ind,
  m,
  n_kernel_param,
  opt_method,
  l_b,
  r_b,
  Jitter
)
}
\description{
Master function for optimizing priors (X_ind, kernel_param)
}
