% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_sgp.R
\name{pois_sgp_update_prior_kernel}
\alias{pois_sgp_update_prior_kernel}
\title{Function for optimize kernel parameters in the prior}
\usage{
pois_sgp_update_prior_kernel(
  X,
  y,
  sc,
  X_ind,
  kernel_param,
  kernel_func,
  post_mean,
  V,
  mu,
  Jitter,
  d_mat_nm,
  d_mat_mm
)
}
\description{
Function for optimize kernel parameters in the prior
}
