% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_sgp.R
\name{pois_sgp_update_V}
\alias{pois_sgp_update_V}
\title{Function for updating posterior var}
\usage{
pois_sgp_update_V(
  X,
  y,
  sc,
  X_ind,
  kernel_param,
  kernel_func,
  post_mean,
  V,
  mu,
  maxiter = 100,
  tol = 1e-05,
  Jitter,
  d_mat_nm,
  d_mat_mm
)
}
\description{
Function for updating posterior var
}
