% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_sgp.R
\name{pois_sgp_update_variational_param}
\alias{pois_sgp_update_variational_param}
\title{Master function for optimizing variational parameters (Kernel_param, mu)}
\usage{
pois_sgp_update_variational_param(
  X,
  y,
  sc,
  X_ind,
  kernel_param,
  kernel_func,
  post_mean,
  V,
  mu,
  fix_kernel_param,
  fix_mu,
  m,
  opt_method,
  Jitter,
  d_mat_nm,
  d_mat_mm,
  verbose
)
}
\description{
Master function for optimizing variational parameters (Kernel_param, mu)
}
