% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgp.R
\name{RBFkernel}
\alias{RBFkernel}
\title{RBF Kernel}
\usage{
RBFkernel(X1, X2, kernel_param, eps = NULL, diagonal = F, d_mat = NULL)
}
\arguments{
\item{d_mat}{output of outer(X1,X2,"-")}
}
\description{
RBF Kernel
}
