% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflect.R
\name{reflect}
\alias{reflect}
\title{A funtion to reflect data}
\usage{
reflect(x, direct = "left", len)
}
\arguments{
\item{x:}{data vector}

\item{direct:}{left, right or both direction}

\item{len:}{the length to reflect}
}
\value{
reflected vector
}
\description{
A funtion to reflect data
}
