% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmad.R
\name{rmad}
\alias{rmad}
\title{Running mad for estimating heteroscedastic variance}
\usage{
rmad(x, filter.number = 1, family = "DaubExPhase")
}
\arguments{
\item{x:}{a vector of data}

\item{filter.number:}{as in smash.gaus}

\item{family:}{as n smash.gaus}
}
\value{
a vector of estimated sd
}
\description{
Running mad for estimating heteroscedastic variance
}
