% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_smooth.R
\name{robust_smooth}
\alias{robust_smooth}
\title{empirical Bayes robust smoothing}
\usage{
robust_smooth(
  y,
  sigma2,
  mu_init = NULL,
  ebnm_control = list(),
  tol = 1e-05,
  maxiter = 100
)
}
\description{
empirical Bayes robust smoothing
}
