% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgp_torch.R
\name{sgp_torch}
\alias{sgp_torch}
\title{sparse GP via torch R}
\usage{
sgp_torch(
  y,
  X = NULL,
  X_ind = NULL,
  m = 10,
  kernel_func = Maternkernel_torch,
  kernel_param = c(0, 0),
  mu = NULL,
  s2 = NULL,
  sigma2 = NULL,
  opt_method = "L-BFGS-B",
  fix_X_ind = T,
  fix_sigma2 = F,
  Jitter = 1e-05
)
}
\description{
sparse GP via torch R
}
