% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma_est.R
\name{sigma_est}
\alias{sigma_est}
\title{A function estimate \sigma^2 using method 4}
\usage{
sigma_est(x, mu, st, method)
}
\arguments{
\item{x:}{data}

\item{method:}{'smashc', 'moment', 'wls', 'ols'}
}
\value{
estimated sd
}
\description{
A function estimate \sigma^2 using method 4
}
