% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simu_study_poisson_smooth}
\alias{simu_study_poisson_smooth}
\title{compare methods for smoothing Poisson sequence}
\usage{
simu_study_poisson_smooth(
  simdata,
  save_data = TRUE,
  method_list = c("vst_nug", "vst_smooth", "lik_exp_mle_nug", "lik_exp_mle_smooth",
    "lik_exp_smashpoi_nug", "lik_exp_smashpoi_smooth", "lik_exp_iter_homo",
    "lik_exp_iter_hetero", "split_runmed_dwt", "split_runmed_ndwt",
    "split_smoothgaus_dwt", "split_smoothgaus_ndwt", "smash", "smash_two_step_homo",
    "smash_two_step_hetero"),
  smoother = "smash",
  n_cores = 20,
  filter.number = 1,
  family = "DaubExPhase",
  maxiter = 100
)
}
\description{
compare methods for smoothing Poisson sequence
}
