% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_study_s.R
\name{simu_study_s}
\alias{simu_study_s}
\title{Test the sigma_est function(estimate sigma)}
\usage{
simu_study_s(mu, st, sigma = 1, var_est = "default", seed = 1234)
}
\value{
a data frame of sigma hat from each method
}
\description{
Test the sigma_est function(estimate sigma)
}
