% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_study_sst.R
\name{simu_study_sst}
\alias{simu_study_sst}
\title{Function to do comparisons among methods(total sd)}
\usage{
simu_study_sst(mu, st, sigma, niter = 100, seed = 12345,
  var_est = "rmad", mu.out = T)
}
\arguments{
\item{mu:}{true mean}

\item{sigma:tue}{sigma}

\item{mu.out:}{whether compare mu.est}
}
\value{
a list of values
}
\description{
Function to do comparisons among methods(total sd)
}
