% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_study_var.R
\name{simu_study_var}
\alias{simu_study_var}
\title{Function to do comparisons among methods}
\usage{
simu_study_var(mu, st, sigma, niter = 100, seed = 12345,
  var_est = "smash")
}
\arguments{
\item{est:}{'s' or 'sst'. 's':sigma only. 'sst': sigma^2+s_t^2}

\item{mu:}{true mean}

\item{sigma:tue}{sigma}
}
\value{
a list of values
}
\description{
Function to do comparisons among methods
}
