% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smash_gen.R
\name{smash_gen}
\alias{smash_gen}
\title{A more general verison of smashgen that works for different data type and variance settings}
\usage{
smash_gen(x, X = NULL, sigma = NULL, ntri = NULL,
  z_var_est = "rmad", wave_family = "DaubExPhase", ashp = TRUE,
  verbose = FALSE, robust = FALSE, filter.number = 1, niter = 1,
  tol = 0.01, dist_family = "poisson", y_var_est = "smashu",
  re_all = FALSE, log_scale = FALSE)
}
\arguments{
\item{x:}{a vector of observations}

\item{X:}{covariates}

\item{sigma:}{standard deviations, scalar. If not known, can be estimated.}

\item{wave_family:}{choice of wavelet basis to be used, as in wavethresh.}

\item{niter:}{number of iterations for IRLS}

\item{tol:}{tolerance of the criterion to stop the iterations}

\item{ashp:}{whether expand around ash posterior mean}

\item{robust:}{whether set the highest resolution wavelet coeffs to 0}

\item{verbose:}{whether print out the number of iterations to converge.}

\item{dist_family:}{family of ditribution: 'poisson','binomial','poi_binom'.}

\item{ntri:}{Binomial number of trials}

\item{y_var_est:}{method to estimate \sigma or \sigma^2+st^2 if \sigma is unknown.}

\item{re_all:}{whether return all estimates including mean and standard deviation(\sigma).}

\item{z_var_est:}{method to estimate variance of intermidiate variable z: 'rmad', 'smash', 'default'}

\item{k:}{parameter in huber m estimator}

\item{log_scale:}{whether return the estimate in log scale(for poisson data only)}
}
\value{
estimated mean or (estimated mean and nuggest effect)
}
\description{
A more general verison of smashgen that works for different data type and variance settings
}
