% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smash_gen_lite.R
\name{smash_gen_lite}
\alias{smash_gen_lite}
\title{A light version for smashgen, to make it fast.}
\usage{
smash_gen_lite(x, sigma = NULL, ntri = NULL, wave_family = "DaubExPhase",
  filter.number = 1, dist_family = "poisson", log_scale = FALSE)
}
\description{
A light version for smashgen, to make it fast.
}
