% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smash_gen_pois_iterative.R
\name{smash_gen_pois_iterative}
\alias{smash_gen_pois_iterative}
\title{Smooth Poisson sequence, accounting for nugget effect, iterative pseudo-likelihood}
\usage{
smash_gen_pois_iterative(
  x,
  s = 1,
  nug.init = NULL,
  nugget_type = "heteroskedastic",
  lik_expan_init = "mle",
  smoother = "smash",
  ash_pm_init_for0 = TRUE,
  eps = "estimate",
  filter.number = 1,
  family = "DaubExPhase",
  maxiter = 10,
  tol = 0.001
)
}
\arguments{
\item{x}{observed Poisson sequence}

\item{s}{Scale factor for Poisson observations: y~Pois(scale*lambda), can be a vector.}

\item{nug.init}{init value of nugget effect, either a scalar or NULL}

\item{nugget_type}{'homoskedastic' or 'heteroskedastic'}

\item{lik_expan_init}{if transformation='lik_expan', where to expand it? Can be mle_pm0, or smash_poi}

\item{smoother}{smoothing method for Gaussian sequence, either 'smash' or 'ti.thresh'. When n is large, ti.thresh is much faster}

\item{eps}{If choose lik_expansion, if x=0, set x = x + eps. Either input a numerical value or 'estimate'. If estimate, eps = sum(x==1)/sum(x<=1)}

\item{filter.number, family}{wavelet basis, see wavethresh package for more details}

\item{maxiter}{max iterations for estimating nugget effect}

\item{tol}{tolerance to stop iterations.}

\item{robust}{whether perform robust wavelet regression}

\item{robust.q}{quantile to determine outliers}

\item{ash.pm}{If choose lik_expansion, whether use ash posterior mean approximation if x=0. If not x = x+eps.}
}
\value{
estimated smoothed lambda, estimated nugget effect.
}
\description{
Smooth Poisson sequence, accounting for nugget effect, iterative pseudo-likelihood
}
