% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smash_gen_x_lite.R
\name{smash_gen_x_lite}
\alias{smash_gen_x_lite}
\title{A light version for smashgen with covariate X, to make it fast.}
\usage{
smash_gen_x_lite(x, X, sigma = NULL, ntri = NULL,
  wave_family = "DaubExPhase", filter.number = 1, dist_family = "poisson",
  log_scale = FALSE, homo = F)
}
\description{
A light version for smashgen with covariate X, to make it fast.
}
