% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smash_two_step.R
\name{smash_two_step}
\alias{smash_two_step}
\title{A two step method for smoothing Poisson seq}
\usage{
smash_two_step(x, homoskedastic = FALSE)
}
\arguments{
\item{x}{data vector}
}
\description{
This method first fit a smash poisson to the seq, then fit smash.gaus to the log estimation.
}
