% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sst_est.R
\name{sst_est}
\alias{sst_est}
\title{Function to estimate \sqrt(\sigma^2+s_t^2) together}
\usage{
sst_est(x, method, filter.number = 1, family = "DaubExPhase")
}
\arguments{
\item{x:}{normal data}

\item{method:}{rmad or smashu}
}
\value{
estimate standard deviation \sqrt(\sigma^2+s_t^2)
}
\description{
Function to estimate \sqrt(\sigma^2+s_t^2) together
}
