% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_gen.R
\name{tf.gen}
\alias{tf.gen}
\title{Trend filtering allow covariates}
\usage{
tf.gen(x, y, ord = 1, lambda = NULL)
}
\arguments{
\item{x:}{covariates, n by p}

\item{y:}{a vector of data}

\item{ord:}{order of trend filter}

\item{lambda:}{reg param, default to be sqrt(n*log(p))}
}
\value{
estimtaed mu and beta
}
\description{
Trend filtering allow covariates
}
