% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_smash_gen.R
\name{update_smash_gen}
\alias{update_smash_gen}
\title{Update function for smashgen algo}
\usage{
update_smash_gen(mu.hat, x, ntri, dist_family)
}
\value{
mt,st,yt
}
\description{
Update function for smashgen algo
}
