% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_sgp.R
\name{ebpm_pois_sgp}
\alias{ebpm_pois_sgp}
\title{Poisson SGP}
\usage{
ebpm_pois_sgp(
  x,
  s = 1,
  g_init = NULL,
  fix_g = FALSE,
  q_init = NULL,
  control = list()
)
}
\arguments{
\item{x}{A vector of Poisson observations.}

\item{s}{A vector of scaling factors for Poisson observations}

\item{g_init}{a list of prior parameters: X_ind, kernel_param, mu, kernel_func}

\item{fix_g}{a boolean or a vector of boolean of length 3. If TRUE, set the corresponding prior at g_init.}

\item{q_init}{a list of init values of posterior: mean, var}

\item{control}{a list of other parameters passed to the `pois_sgp` function.}
}
\description{
Poisson SGP
}
