% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebps.R
\name{ebps}
\alias{ebps}
\title{empirical Bayes Poisson smooting}
\usage{
ebps(
  x,
  s = NULL,
  g_init = NULL,
  q_init = NULL,
  init_control = list(),
  general_control = list(),
  smooth_control = list()
)
}
\arguments{
\item{x}{data vector}

\item{maxiter, tol}{max iteration and tolerance for stopping it.}

\item{m_init, sigma2_init}{initial values of latent variable and nugget effect.}

\item{smooth_init}{init of smooth function.}

\item{wave_trans}{dwt or ndwt. If ndwt, stopping criteria cannot be `objabs`}

\item{ndwt_method}{if wave_trans is ndwt, either use `smash` or `ti.thresh`. When n is large, `ti.thresh` is much faster.}

\item{convergence_criteria}{'objabs' for absolute diff in ELBO, 'nugabs' for absolute diff in nugget effect}

\item{warmstart}{whether warmstart of dwt smoother}
}
\description{
empirical Bayes Poisson smooting
}
\details{
The problem is
\deqn{x_i\sim Poisson(\lambda_i,}
\deqn{\lambda_i = \exp(\mu_i)),}
\deqn{\mu_i\sim N(b_i,\sigma^2),}
\deqn{\b_i\sim g(.).}
}
\examples{
set.seed(12345)
n=2^9
sigma=0.5
mu=c(rep(0.3,n/4), rep(3, n/4), rep(10, n/4), rep(0.3, n/4))
x = rpois(n,exp(log(mu)+rnorm(n,sd=sigma)))
fit = pois_smooth_split(x,maxiter=30)
plot(x,col='grey80')
lines(fit$posterior$mean_smooth)
fit$sigma2
plot(fit$elbo_trace)
}
