% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_smash_gen.R
\name{init_smash_gen}
\alias{init_smash_gen}
\title{Initialize the input of smashgen algorithm}
\usage{
init_smash_gen(x, ntri, ashp, dist_family)
}
\arguments{
\item{x:}{input data}

\item{ntri:}{Binomial number of trials}

\item{ashp:}{whether apply ash firstly}

\item{dist_family:}{family of ditribution: 'poisson','binomial','poi_binom'.}
}
\value{
initilized m,s,y
}
\description{
Initialize the input of smashgen algorithm
}
