% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_sgp.R
\name{pois_sgp}
\alias{pois_sgp}
\title{Sparse Gaussian process for Poisson data via variational inference}
\usage{
pois_sgp(
  y,
  X = NULL,
  X_ind = NULL,
  m = 30,
  post_mean = NULL,
  V = NULL,
  kernel_func = Maternkernel,
  kernel_param = NULL,
  mu = NULL,
  sc = NULL,
  opt_method = "L-BFGS-B",
  fix_X_ind = F,
  fix_kernel_param = F,
  fix_mu = F,
  maxiter = 100,
  tol = 1e-05,
  maxiter_mean = 100,
  tol_mean = 1e-05,
  maxiter_V = 100,
  tol_V = 1e-05,
  l_b = -Inf,
  r_b = Inf,
  Jitter = 1e-05,
  verbose = T,
  printevery = 1
)
}
\arguments{
\item{y}{count vector}

\item{X, X_ind}{grids, and inducing points}

\item{m}{number of inducing points}

\item{post_mean, V}{posterior mean and var of f at inducing point}

\item{kernel_func, kernel_param}{functions and their parameters}

\item{sc}{scaling scalar or vectors}
}
\description{
Sparse Gaussian process for Poisson data via variational inference
}
