% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_sgp.R
\name{pois_sgp_get_posterior}
\alias{pois_sgp_get_posterior}
\title{Obtain full posterior distribution}
\usage{
pois_sgp_get_posterior(
  X,
  y,
  sc,
  X_ind,
  kernel_param,
  kernel_func,
  post_mean,
  V,
  mu,
  Jitter,
  d_mat_nm,
  d_mat_mm
)
}
\description{
Obtain full posterior distribution
}
