% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_smooth_split.R
\name{pois_smooth_split}
\alias{pois_smooth_split}
\title{Smooth over-dispersed Poisson sequence via splitting method}
\usage{
pois_smooth_split(
  x,
  s = NULL,
  Eb_init = "runmed",
  sigma2_init = NULL,
  est_sigma2 = TRUE,
  maxiter = 100,
  tol = 1e-05,
  filter.number = 1,
  family = "DaubExPhase",
  wave_trans = "dwt",
  ndwt_method = "smash",
  verbose = FALSE,
  printevery = 10,
  ebnm_params = list(mode = 0),
  ashparam = list(),
  optim_method = "L-BFGS-B",
  convergence_criteria = "objabs",
  W = NULL
)
}
\arguments{
\item{x}{data vector}

\item{Eb_init, sigma2_init}{initial values of smooth mean and nugget effect.}

\item{maxiter, tol}{max iteration and tolerance for stopping it.}

\item{wave_trans}{dwt or ndwt. If ndwt, stopping criteria cannot be `objabs`}

\item{ndwt_method}{if wave_trans is ndwt, either use `smash` or `ti.thresh`. When n is large, `ti.thresh` is much faster.}

\item{convergence_criteria}{'objabs' for absolute diff in ELBO, 'nugabs' for absolute diff in nugget effect}
}
\description{
Smooth over-dispersed Poisson sequence via splitting method
}
\details{
The problem is
\deqn{x_i\sim Poisson(\lambda_i,}
\deqn{\lambda_i = \exp(\mu_i)),}
\deqn{\mu_i\sim N(b_i,\sigma^2),}
\deqn{\b_i\sim g(.).}
}
\examples{
set.seed(12345)
n=2^9
sigma=0.5
mu=c(rep(0.3,n/4), rep(3, n/4), rep(10, n/4), rep(0.3, n/4))
x = rpois(n,exp(log(mu)+rnorm(n,sd=sigma)))
fit = pois_smooth_split(x,maxiter=30)
plot(x,col='grey80')
lines(exp(fit$Eb))
fit$sigma2
plot(fit$obj)
}
