% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgp.R
\name{sgp}
\alias{sgp}
\title{Sparse Gaussian Process via VFE}
\usage{
sgp(
  X,
  y,
  X_ind,
  kernel_func = RBFkernel,
  kernel_param = c(0, 0),
  mu = NULL,
  s2 = NULL,
  sigma2 = 1,
  opt_method = "L-BFGS-B",
  fix_x_ind = F,
  fix_kernel_param = F,
  fix_sigma2 = F,
  fix_mu = F,
  l_b = -Inf,
  r_b = Inf
)
}
\description{
Sparse Gaussian Process via VFE
}
