% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{sim_data_smooth}
\alias{sim_data_smooth}
\title{generate Poisson sequence data}
\usage{
sim_data_smooth(
  n_simu,
  n = 2^9,
  snr = 3,
  count_size,
  smooth_func = "blocks",
  seed = 12345
)
}
\arguments{
\item{n_simu}{number of simulation reps}

\item{n}{length of Poisson seq}

\item{snr}{signal  to noise ratio}

\item{count_size}{max of exp(b)}

\item{smooth_func}{cblocks, sblocks, heavi, angles, bursts, spike}
}
\value{
a list of
 \item{X:}{data}
 \item{L:}{mean}
 \item{other imputs:}{sigma2,snr, ...}
}
\description{
generate Poisson sequence data
}
