% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simu_study_poisson_smooth}
\alias{simu_study_poisson_smooth}
\title{compare methods for smoothing Poisson sequence}
\usage{
simu_study_poisson_smooth(
  simdata,
  save_data = TRUE,
  method_list = c("vst", "lik_exp", "split_dwt", "split_ndwt", "smash", "smash_two_step"),
  n_cores = 10,
  filter.number = 1,
  family = "DaubExPhase",
  maxiter = 100
)
}
\description{
compare methods for smoothing Poisson sequence
}
