% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simu_study_poisson_smooth}
\alias{simu_study_poisson_smooth}
\title{compare methods for smoothing Poisson sequence}
\usage{
simu_study_poisson_smooth(
  simdata,
  save_data = TRUE,
  method_list = c("vst_nug", "vst_nug_top", "vst_smooth", "lik_exp_mle_nug",
    "lik_exp_mle_nug_top", "lik_exp_mle_smooth", "lik_exp_smashpoi_nug",
    "lik_exp_smashpoi_nug_top", "lik_exp_smashpoi_smooth", "split_dwt_true",
    "split_runmed_dwt", "split_runmed_dwt_fix_nug", "split_smashpoi_dwt",
    "split_smashpoi_dwt_fix_nug", "split_log1px_dwt", "split_log1px_dwt_fix_nug",
    "split_runmed_ndwt", "split_runmed_ndwt_top", "split_runmed_ndwt_fix_nug",
    "split_log1px_ndwt", "split_log1px_ndwt_fix_nug", "split_smashpoi_ndwt",
    "split_smashpoi_ndwt_fix_nug", 
     "split_ndwt_true", "smash_two_step_homo",
    "smash_two_step_hetero", "smash"),
  smoother = "smash",
  n_cores = 20,
  filter.number = 1,
  family = "DaubExPhase",
  nug.est.limit = 0.2,
  maxiter = 100
)
}
\description{
compare methods for smoothing Poisson sequence
}
