% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smash_gaus_gen.R
\name{smash.gaus.gen}
\alias{smash.gaus.gen}
\title{function to estimate both mu and \sigma}
\usage{
smash.gaus.gen(x, st, family = "DaubExPhase", niters = 2,
  method = "smashc")
}
\arguments{
\item{st:}{known variance}

\item{family:}{wavelet basis}

\item{niters:}{number of iterations to estimate \sigma and mu}

\item{method:}{'smashc', 'ols' amd 'moment'}
}
\value{
estimated mean and \sigma
}
\description{
function to estimate both mu and \sigma
}
