% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smash_gaus_gen.R
\name{smash.gaus.gen}
\alias{smash.gaus.gen}
\title{function to estimate both mu and \sigma}
\usage{
smash.gaus.gen(x, st, y_var_est = "mle", z_var_est = "smash",
  family = "DaubExPhase", filter.number = 1, niters = 2, k = NULL)
}
\arguments{
\item{st:}{known variance}

\item{family:}{wavelet basis}

\item{filter.number:}{as in smash.gaus}

\item{niters:}{number of iterations to estimate \sigma and mu}

\item{y_var_est:}{'mle', 'moment', 'eb', 'huber','wls'}

\item{z_var_est:}{method to estimate variance: 'rmad', 'smash', 'default'}

\item{k:}{parameter in huber m estimator}
}
\value{
estimated mean and \sigma
}
\description{
function to estimate both mu and \sigma
}
