% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smash_gaus_x.R
\name{smash.gaus.x}
\alias{smash.gaus.x}
\title{smash gaus to deal with covariate effect.}
\usage{
smash.gaus.x(X, y, filter.number = 1, family = "DaubExPhase", homo = T)
}
\arguments{
\item{X:}{covariate, n by p}

\item{y:}{a vector of data}

\item{homo:}{wether homoscedastic or heteroscedatic variance.}
}
\value{
a list: mu.hat, beta.hat, se.hat
}
\description{
smash gaus to deal with covariate effect.
}
