% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smash_gen.R
\name{smash_gen}
\alias{smash_gen}
\title{A more general verison of smashgen that works for different data type and variance settings}
\usage{
smash_gen(x, sigma = NULL, ntri = NULL, wave_family = "DaubExPhase",
  ashp = TRUE, verbose = FALSE, robust = FALSE, niter = 1, tol = 0.01,
  dist_family = "poisson", method = "smashu", reall = FALSE)
}
\arguments{
\item{x:}{a vector of observations}

\item{sigma:}{standard deviations, scalar. If not known, can be estimated.}

\item{wave_family:}{choice of wavelet basis to be used, as in wavethresh.}

\item{niter:}{number of iterations for IRLS}

\item{tol:}{tolerance of the criterion to stop the iterations}

\item{ashp:}{whether expand around ash posterior mean}

\item{robust:}{whether set the highest resolution wavelet coeffs to 0}

\item{verbose:}{whether print out the number of iterations to converge.}

\item{dist_family:}{family of ditribution: 'poisson','binomial'.}

\item{ntri:}{Binomial number of trials}

\item{method:}{method to estimate \sigma or \sigma^2+st^2 if \sigma is unknown.}

\item{reall:}{whether return all estimates including mean and standard deviation(\sigma, only when method='smashc','ols' or 'moment').}
}
\value{
estimated mean
}
\description{
A more general verison of smashgen that works for different data type and variance settings
}
